Increase CoreAudioCaptureSource capture verify timer to handle the case of device change https://bugs.webkit.org/show_bug.cgi?id=233250
Reviewed by Eric Carlson.

Previously, our capture verify timer was starting with 10 seconds to handle slow start up times, and decreasing to 2 seconds to be snappy.
We are now sticking to 10 seconds to handle the case of changing of capture device while capturing.
This happens on iOS for instance when going from builtin microphone to BT headset microphone.

Manually tested.

* platform/mediastream/mac/CoreAudioCaptureSource.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285927 268f45cc-cd09-0410-ab3c-d52691b4dbfc