[WebAuthn] Add SPI for makeCredential / getAssertion using clientDataHash https://bugs.webkit.org/show_bug.cgi?id=233216 <rdar://problem/85476386>
Reviewed by Brent Fulgham.

In order to avoid needing to make and coordinate changes to ASC to support new fields or changes
within ClientDataJSON and to maintain a single source of truth, calls to ASC from WebKit
will contain a precomputed ClientDataHash. This change creates new SPIs that will be called
from ASC using the ClientDataHash.

* UIProcess/API/Cocoa/_WKAuthenticatorResponse.h:
* UIProcess/API/Cocoa/_WKWebAuthenticationPanel.mm:
(-[_WKWebAuthenticationPanel makeCredentialWithClientDataHash:options:completionHandler:]):
(-[_WKWebAuthenticationPanel getAssertionWithClientDataHash:options:completionHandler:]):
New functions to take in ClientDataHash instead of the data needed to construct it.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285965 268f45cc-cd09-0410-ab3c-d52691b4dbfc