[css-grid] Transfer size for grid item with an aspect-ratio and stretch alignment against the definite row https://bugs.webkit.org/show_bug.cgi?id=232987
Reviewed by Javier Fernandez.

Source/WebCore:

For a grid item with an aspect-ratio, if it has stretch alignment against the definite row, it should
try and resolve it if possible and transfer this size into the inline direction for the min/max content
size. For the case that the grid width is content sized, we need to update the width before laying out
the grid items. Since the min-content contribution of the grid item has changed based on the row sizes
calculated in step 2 of sizing algorithm, we also need to repeat the sizing algorithm steps to update
the width of the track sizes.

* rendering/GridLayoutFunctions.cpp:
(WebCore::GridLayoutFunctions::isAspectRatioBlockSizeDependentChild):
* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::updateGridAreaForAspectRatioItems):
* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::needsPreferredWidthsRecalculation const):

LayoutTests:

* TestExpectations: unskip the test that is now passing.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285987 268f45cc-cd09-0410-ab3c-d52691b4dbfc