Fix assertion added in r285592 https://bugs.webkit.org/show_bug.cgi?id=233373 JSTests:
Reviewed by Keith Miller.

* stress/validate-assert-uid-is-not-index-or-it-is-symbol.js: Added.
(main.v11):
(main.v29):
(main):

Source/JavaScriptCore:

rdar://85451012

Reviewed by Keith Miller.

The assertion added in r285592 should not apply to Symbols. This patch
fixes that error. We don't care if a Symbol can be parsed as an index
since the string value in a Symbol is just its description, not the
actual property.

* dfg/DFGValidate.cpp:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286070 268f45cc-cd09-0410-ab3c-d52691b4dbfc