[GTK] Layout Test http/tests/media/video-play-waiting.html is timing out https://bugs.webkit.org/show_bug.cgi?id=142489
Reviewed by Philippe Normand.

The test was failing with a timeout even after properly calling TestRunner.notifyDone()
internally because there was still an injectedBundle.topLoadingFrame() in InjectedBundlePage,
which means that the web page hadn't completely finished to be downloaded. This was caused
by the stalled video download that the test wants to exercise.

The solution is to set the video src to empty in order to convince TestRunner that the page
has finished downloading.

* http/tests/media/video-play-waiting.html: Set the video src to empty.
* platform/mac/TestExpectations: The test now just fails (for reasons unrelated to this fix) on mac instead of timing out.
* platform/glib/TestExpectations: Unskipped test.
* platform/win/TestExpectations: Unskipped test.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286165 268f45cc-cd09-0410-ab3c-d52691b4dbfc