[css-flexbox] Do not shrink tables bellow their intrinsic sizes https://bugs.webkit.org/show_bug.cgi?id=232383
Reviewed by Manuel Rego Casasnovas.

Source/WebCore:

Flex layout algorithm uses the flex items min|max sizes to compute the hypothetical main size which
will be used to clamp the size of the items when flexed. In the case of tables as flex items, we were
using the min-size straight away whenever that was definite. However we cannot do that because tables
must not shrink bellow their intrinsic sizes. That's why we should select the maximum of the intrinsic
size and the min-size.

This fixes a WPT test covering this case.

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::computeFlexItemMinMaxSizes): Do not select a min-size smaller than the
table intrinsic size in the main axis.

LayoutTests:

* TestExpectations: Unskip table-as-item-fixed-min-width-3.html which works fine now.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286207 268f45cc-cd09-0410-ab3c-d52691b4dbfc