Add a basic heuristic for sizing text in image overlay blocks https://bugs.webkit.org/show_bug.cgi?id=233343 rdar://85755755
Reviewed by Tim Horton.

Source/WebCore:

Address a FIXME by implementing a simple heuristic for sizing text in image overlay blocks. Each image overlay
block has a predetermined target width and height; currently, we hard-code the font size in each of these blocks
to be 80% of the target height. While this works in some cases, it also causes content in these blocks to
overflow the container if there's too much text.

Avoid this problem by finding the largest font size, such that text in each container fits within the target
bounds without overflowing vertically (we don't need to worry about horizontal overflow because horizontally
overflowing text will simply wrap to the next line).

Because of line wrapping, it's difficult to analytically compute this font size, so I'm instead using binary
search (up to a fixed number of iterations) to converge on this largest size that avoids exceeding the target
height. See below for more details.

Test: fast/images/text-recognition/image-overlay-blocks.html

* dom/ImageOverlay.cpp:
(WebCore::ImageOverlay::updateWithTextRecognitionResult):

Implement the main logic of the binary search. Establish a goal of ending up with text that fills at least 90%
of the available space in the container (with a max of just slightly over 100%, so that we stop early if we
happen to only barely exceed the target height). Then, starting with a font size that's 80% of the target
height, converge on a font size between 0% and 100% of the target height that achieves this goal. In practice,
most of the injected blocks of text will stop after a single iteration (due to the "80%-of-height" font size
being sufficient to mostly fill available vertical space).

Each FontSizeAdjustmentState here represents a single block container. At the beginning of each iteration, we
ensure that layout is up to date, and set the font size to be between the current min and max value (scaling by
target height to get font size). Using `linesBoundingBox()`, we compute the height of the rendered text relative
to its parent (which we know a-priori to be as tall as the target height). If this height is at least 90% of the
target height, we finish and remove the font size adjustment state from the vector of pending font adjustments;
otherwise, we adjust the min and max to either increase or decrease the font size as needed, and continue the
loop. We finally bail after an arbitrarily-chosen 10 iterations, if the text in the container is still too
small.

* testing/Internals.cpp:
(WebCore::Internals::installImageOverlay):

Augment this testing-only internals hook so that we can inject arbitrary image overlay block content inside
image overlays.

* testing/Internals.h:
* testing/Internals.idl:

LayoutTests:

Add a new layout test. See WebCore/ChangeLog for more details.

* fast/images/text-recognition/image-overlay-blocks-expected.html: Added.
* fast/images/text-recognition/image-overlay-blocks.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286265 268f45cc-cd09-0410-ab3c-d52691b4dbfc