[Webpage translation] Avoid translating text inside image overlays https://bugs.webkit.org/show_bug.cgi?id=233225 rdar://85713734
Reviewed by Aditya Keerthi.

Source/WebCore:

Add `translate="no"` to image overlay content; text in image overlays is meant to visually match recognized text
in the image, and should never be modified as a part of webpage translation.

Test: fast/images/text-recognition/image-overlay-do-not-translate.html

* dom/ImageOverlay.cpp:
(WebCore::ImageOverlay::updateSubtree):

LayoutTests:

Add a test to verify that `translate` is `false` on the image overlay root container element.

* fast/images/text-recognition/image-overlay-do-not-translate-expected.txt: Added.
* fast/images/text-recognition/image-overlay-do-not-translate.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286282 268f45cc-cd09-0410-ab3c-d52691b4dbfc