[:has() pseudo-class] Cache :has() failures for subtrees https://bugs.webkit.org/show_bug.cgi?id=233631
Reviewed by Simon Fraser.

Add a temporary (single style resolution/invalidation scoped) cache that remembers whether
a given :has() argument has any matches in a subtree. This avoids repeated tree walks in
presence of very generic descendant matching :has() selectors.

* css/SelectorChecker.cpp:
(WebCore::SelectorChecker::checkOne const):

This optimization can be used if the :has() not matching wasn't about the scope.

(WebCore::SelectorChecker::matchHasPseudoClass const):

When we traverse a subtree and find no matches for a has selector add a cache entry saying so.
Before traversing test if any of the ancestors elements already have a cache entry saying there
are no matches in this subtree for this selector.

* css/SelectorChecker.h:
* style/ElementRuleCollector.cpp:
(WebCore::Style::ElementRuleCollector::ruleMatches):
* style/SelectorMatchingState.h:
(WebCore::Style::makeHasPseudoClassCacheKey):

Add the cache to SelectorMatchingState. It is available when selector filter is.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286302 268f45cc-cd09-0410-ab3c-d52691b4dbfc