Correct serialization error in _WKApplicationManifestIcon https://bugs.webkit.org/show_bug.cgi?id=233602 <rdar://problem/85838843>
Reviewed by Devin Rousso.

Source/WebKit:

The 'initWithCoder' method had two mistakes: (1) It attempted to decode the 'purposes'
member as an NSArray of NSString, rather than NSNumber, and (2) it attempted to decode
the 'src' member as an NSString, when it is in fact an NSURL.

These errors caused serialization to fail (since NSSecureCoding correctly rejected the
serialization).

This patch corrects both errors and adds a new test to exercise this code path.

Tested by a new TestWebKitAPI Test: ApplicationManifest.IconCoding

* UIProcess/API/Cocoa/_WKApplicationManifest.mm:
(-[_WKApplicationManifestIcon initWithCoder:]): Correct decoding logic.

Tools:

This patch adds a new test "ApplicationManifest.IconCoding" to exercise the
NSCoder code paths.

* TestWebKitAPI/Tests/WebKitCocoa/ApplicationManifest.mm:
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286311 268f45cc-cd09-0410-ab3c-d52691b4dbfc