nullptr deref in InbandChapterTrackPrivateAVFObjC::processChapters https://bugs.webkit.org/show_bug.cgi?id=233605
Patch by Gabriel Nava Marino <gnavamarino@apple.com> on 2021-11-30
Reviewed by Eric Carlson.

Source/WebCore:

The createChapterCue lambda gets dispatched onto the main thread. When it executes,
the client might no longer be valid, so we are adding a nullptr check at the start
of the lambda, as is done at the start of the method.

Test: media/track/track-in-band-chapters-invalid-client-crash.html

* platform/graphics/avfoundation/objc/InbandChapterTrackPrivateAVFObjC.mm:
(WebCore::InbandChapterTrackPrivateAVFObjC::processChapters):

LayoutTests:

* media/track/track-in-band-chapters-invalid-client-crash-expected.txt: Added.
* media/track/track-in-band-chapters-invalid-client-crash.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286327 268f45cc-cd09-0410-ab3c-d52691b4dbfc