Reuse navigation preload if service worker is fetching the corresponding navigation request https://bugs.webkit.org/show_bug.cgi?id=233490
Reviewed by Chris Dumez.

Source/WebCore:

Store the fetch identifier in FetchRequest as this might be used as the navigation preload identifier.
We set it when creating a FetchRequest for firing a fetch event.
This identifier is kept in case of request cloning or when creating a request as part of calling fetch.
It is used to set the navigationPreloadIdentifier in ResourceLoaderOptions that is piped to WebLoaderStrategy.

Test: http/wpt/service-workers/fetch-service-worker-preload.https.html

* Modules/fetch/FetchLoader.cpp:
* Modules/fetch/FetchRequest.cpp:
* Modules/fetch/FetchRequest.h:
* loader/ResourceLoaderOptions.h:
* loader/ThreadableLoader.cpp:
* workers/service/context/ServiceWorkerFetch.cpp:
* workers/service/context/ServiceWorkerFetch.h:
* workers/service/context/ServiceWorkerThread.cpp:
* workers/service/context/ServiceWorkerThread.h:
* workers/service/context/ServiceWorkerThreadProxy.cpp:

Source/WebKit:

Store navigation preloader in NetworkSession, keyed by the corresponding fetch identifier.
When a navigation request is coming from fetch with the corresponding navigation preload identifier
we look at ongoing preloads and use it if available to answer the fetch.
To do so, we create a new ServiceWorkerFetchTask that will take the preloader of the initial ServiceWorkerFetchTask.

* NetworkProcess/NetworkResourceLoadParameters.cpp:
* NetworkProcess/NetworkResourceLoadParameters.h:
* NetworkProcess/NetworkSession.cpp:
* NetworkProcess/NetworkSession.h:
* NetworkProcess/ServiceWorker/ServiceWorkerFetchTask.cpp:
* NetworkProcess/ServiceWorker/ServiceWorkerFetchTask.h:
* NetworkProcess/ServiceWorker/WebSWServerConnection.cpp:
* WebProcess/Network/WebLoaderStrategy.cpp:

LayoutTests:

* http/wpt/service-workers/fetch-service-worker-preload-worker.js: Added.
* http/wpt/service-workers/fetch-service-worker-preload.https-expected.txt: Added.
* http/wpt/service-workers/fetch-service-worker-preload.https.html: Added.
* http/wpt/service-workers/resources/fetch-service-worker-preload-script.py: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286361 268f45cc-cd09-0410-ab3c-d52691b4dbfc