[LFC][IFC] Introduce createDisplayBoxesInVisualOrderForContentRuns https://bugs.webkit.org/show_bug.cgi?id=233733
Reviewed by Antti Koivisto.

Revert some of the inline box related bidi changes and start with
a very simple "let's create the display boxes for all _content_ runs in visual order".
This is very similar to InlineDisplayContentBuilder::processNonBidiContent() where we just
loop through the content and construct display boxes.
This is also in preparation for advancing with the inline box case.

* layout/formattingContexts/inline/InlineDisplayContentBuilder.cpp:
(WebCore::Layout::InlineDisplayContentBuilder::processNonBidiContent):
(WebCore::Layout::InlineDisplayContentBuilder::processBidiContent):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286446 268f45cc-cd09-0410-ab3c-d52691b4dbfc