[iOS] Web content process sometimes crashes under WebPage::positionInformation() https://bugs.webkit.org/show_bug.cgi?id=233841 rdar://85917212
Reviewed by Geoffrey Garen.

Source/WebKit:

Add a missing null check in the case where the hit-tested node is null in `WebPage::positionInformation`. This
can happen in a number of ways (one of which is exercised by the new API test).

Test: ImageAnalysisTests.DoNotCrashWhenHitTestingOutsideOfWebView

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::positionInformation):

Tools:

Add a test that exercises hit-testing for Live Text, but induces this crash by hit-testing outside of the bounds
of the web view (such that the hit-tested node ends up as null).

* TestWebKitAPI/Tests/WebKitCocoa/ImageAnalysisTests.mm:
(TestWebKitAPI::swizzledLocationInView):
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286522 268f45cc-cd09-0410-ab3c-d52691b4dbfc