REGRESSION(r286535) [GTK] Fix clean builds after DerivedSources/WebKit2 move https://bugs.webkit.org/show_bug.cgi?id=233881
Reviewed by Don Olmstead.

Build fix. Covered by existing tests

The previous approach of this patch created the links to the header
derived sources into the target webkit2/ folder, which in turn was a
link to <repo>/Source/WebKit/UIProcess/API/gtk. So it ended up
actually creating the links inside the Source directory.

This commit instead links a different webkit2/ folder from the
generated webkit folder for the WebKitGTK/WPE derived sources headers.

* PlatformGTK.cmake:
* PlatformWPE.cmake:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@286585 268f45cc-cd09-0410-ab3c-d52691b4dbfc